+++ apps/api/src/services/boost-award.service.ts
@@ -1,7 +1,9 @@
 import type {
   BoostDiffResult,
   BoostRepository,
-} from "@/database/repositories/boost.repository.js";
+} from "@recallnet/db-schema/repositories/boost";
+import { Transaction } from "@recallnet/db-schema/types";
+
 import type { UserService } from "@/services/user.service.js";
 
 export { BoostAwardService };
