+++ apps/api/src/services/boost.service.ts
@@ -1,7 +1,8 @@
 import {
   type BoostAgentResult,
   BoostRepository,
-} from "@/database/repositories/boost.repository.js";
+} from "@recallnet/db-schema/repositories/boost";
+
 import { serviceLogger } from "@/lib/logger.js";
 import { ApiError } from "@/middleware/errorHandler.js";
 import type { CompetitionService } from "@/services/competition.service.js";
