+++ packages/db-schema/src/repositories/boost.ts
@@ -8,12 +8,21 @@ import { Database } from "../types.js";
 import * as schema from "../voting/defs.js";
 import { SelectAgentBoost, SelectAgentBoostTotal } from "../voting/types.js";
 
+export { BoostRepository, BoostChangeMetaSchema };
+export type {
+  BoostDiffArgs,
+  BoostDiffResult,
+  BoostChangeMeta,
+  BoostAgentArgs,
+  BoostAgentResult,
+};
+
 /** Schema of an optional structured context to attach to each boost change. */
-export const BoostChangeMetaSchema = z.object({
+const BoostChangeMetaSchema = z.object({
   description: z.string().optional(),
 });
 /** Optional structured context to attach to each boost change. */
-export type BoostChangeMeta = z.infer<typeof BoostChangeMetaSchema>;
+type BoostChangeMeta = z.infer<typeof BoostChangeMetaSchema>;
 const DEFAULT_META: BoostChangeMeta = {};
 
 /**
@@ -23,7 +32,7 @@ const DEFAULT_META: BoostChangeMeta = {};
  * - Provide a stable `idemKey` to make the operation retry-safe.
  *   If omitted, a random UUID is generated (non-idempotent across retries).
  */
-export type BoostDiffArgs = {
+type BoostDiffArgs = {
   /** User ID */
   userId: string;
   /** EVM address; will be lowercased before persisting. */
@@ -52,7 +61,7 @@ export type BoostDiffArgs = {
  * - Applied: returns new balance and the created change id.
  * - Idempotent no-op: returns existing balance when the same idemKey was seen.
  */
-export type BoostDiffResult =
+type BoostDiffResult =
   | {
       type: "applied";
       changeId: string;
@@ -61,7 +70,7 @@ export type BoostDiffResult =
     } // change applied
   | { type: "noop"; balance: bigint; idemKey: Uint8Array }; // noop (already applied)
 
-export type BoostAgentArgs = {
+type BoostAgentArgs = {
   userId: string;
   /** EVM address; will be lowercased before persisting. */
   wallet: string;
@@ -79,7 +88,7 @@ export type BoostAgentArgs = {
   idemKey?: Uint8Array;
 };
 
-export type BoostAgentResult =
+type BoostAgentResult =
   | {
       type: "applied";
       agentBoost: SelectAgentBoost;
@@ -152,11 +161,11 @@ export type BoostAgentResult =
  *     Use separate keys; if you must reverse, create a **new** key with a negative/positive
  *     delta as appropriate.
  */
-export class BoostRepository {
+class BoostRepository {
   readonly #db: Database;
 
-  constructor(db: Database) {
-    this.#db = db;
+  constructor(database: Database) {
+    this.#db = database;
   }
 
   /**
@@ -465,7 +474,7 @@ export class BoostRepository {
       competitionId: string;
     },
     tx?: Transaction,
-  ) {
+  ): Promise<bigint> {
     const executor = tx || this.#db;
     const [res] = await executor
       .select()
@@ -483,7 +492,7 @@ export class BoostRepository {
   async agentBoostTotals(
     { competitionId }: { competitionId: string },
     tx?: Transaction,
-  ) {
+  ): Promise<Record<string, bigint>> {
     const executor = tx || this.#db;
     const res = await executor
       .select()
@@ -529,7 +538,7 @@ export class BoostRepository {
   async userBoosts(
     { userId, competitionId }: { userId: string; competitionId: string },
     tx?: Transaction,
-  ) {
+  ): Promise<Record<string, bigint>> {
     const executor = tx || this.#db;
     const res = await executor
       .select({
